requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
